requirements.txt
Flask
flask-pymongo
flask-bcrypt
Flask-JWT-Extended
flask-cors
openai
python-dotenv
